+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 <!-- There is always Unreleased section on the top. Subsections (Add, Changed, Fix, Removed) should be Add as needed. -->
 ## Unreleased
+- Allow generic `ConsumeHandler` function
+    - for `kafkaApplication` (and `partialKafkaApplication`) in `consume` and `consumeFrom`
+- Make some intentionally internal types really `internal`
 
 ## 18.2.0 - 2022-02-11
 - Allow generic `DeriveEvent` function